Teaching-pack-generator-completed/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "repo"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
        "chainlit>=2.5.5",
        "google-generativeai>=0.8.5",
        "google-genai>=0.1.0",
        "loguru>=0.7.3",
        "openpyxl>=3.1.5",
        "pandas>=2.3.0",
        "pydantic-ai>=1.39.0",
        "pypdf2>=3.0.0",
        "python-docx>=1.1.0",
        "python-dotenv>=1.1.0",
        "sentence-transformers>=2.7.0",
        "requests>=2.31.0",
        "fastapi>=0.115.0",
        "uvicorn[standard]>=0.32.0",
        "python-multipart>=0.0.9",
        "sqlalchemy>=2.0.0",
        "alembic>=1.13.0",
        "psycopg2-binary>=2.9.0",
        "python-jose[cryptography]>=3.3.0",
        "passlib[bcrypt]>=1.7.4",
        "pytesseract>=0.3.10",
        "pdf2image>=1.17.0",
        # =====================
        # Job Queue / Worker
        # =====================
        "redis>=5.0.0",
        "rq>=1.16.0",

        # =====================
        # Production Server
        # =====================
        "gunicorn>=21.2.0",

        # =====================
        # HTTP / Assets (used in code)
        # =====================
        "httpx>=0.27.0",
        "pillow>=10.2.0",
        "python-pptx>=0.6.23",
        "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "black>=24.2.0",
    "mypy>=1.8.0",
    "flake8>=7.0.0",
    "promptfoo",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["data*", "handlers*", "llm*", "mcp*", "utils*", "models*"]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi...$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["pymilvus.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
    ".tox",
]
